+++ components/meals/index.ts
@@ -1,5 +1,1 @@
-export { default as PantrySidebar } from './PantrySidebar';
-export { default as ShoppingListSidebar } from './ShoppingListSidebar';
-export { default as RecipesSidebar } from './RecipesSidebar';
-
-
+export { MealsView } from './MealsView';