+++ components/modals/index.ts
@@ -2,4 +2,4 @@ export { ProjectFormModal } from './ProjectFormModal';
 export { ProjectTaskFormModal } from './ProjectTaskFormModal';
 export { TaskFormModal } from './TaskFormModal';
 export { QuickAddTaskModal } from './QuickAddTaskModal'; 
-export { RecipeFormModal } from './RecipeFormModal';
+export { MealFormModal } from './MealFormModal';