+++ components/recipes/index.ts
@@ -1,1 +1,0 @@
-export { RecipesView } from './RecipesView';